loan_process/requirements_no_version.txt
asgiref
certifi
chardet
Django
django-appconf
django-cors-headers
django-cryptography
django-extensions
django-filter
djangorestframework
djangorestframework-simplejwt
Faker
idna
Jinja2
joblib
jsonpatch
jsonpointer
jsonschema
lightgbm
markdown-it-py
MarkupSafe
mdurl
mysqlclient
numpy
pandas
pillow
PyJWT
pyparsing
pyrsistent
python-dateutil
pytz
PyYAML
requests
rich
scikit-learn
scipy
six
sqlparse
threadpoolctl
typing_extensions
tzdata
urllib3
wheel
xgboost